refactor(mcp): drop dead read-only flag, dedup settings/result/redis helpers
Code-quality pass on the MCP read-only work:

- Remove the always-true `read_only` flag. The only caller hardcoded `true`,
  so the field, the start_mcp_server param, and the branch in execute_query
  were dead configurability whose `else` arm silently routed to the writable
  path. The MCP server is now unconditionally read-only. Fix the execute_query
  tool annotation to match (read_only=true, destructive=false).
- Add QueryResult::from_rows / from_error and use them in the four
  execute_query_read_only impls, removing the copy-pasted result shaping.
- Extract the Redis read-only allowlist into database::redis_read_only (with
  its tests), shrinking the oversized redis.rs and isolating the
  security-critical allowlist. Add the missing safe reads it should accept
  (BITCOUNT/BITPOS/GETBIT/BITFIELD_RO/HSTRLEN/SORT_RO/LCS/EXPIRETIME/PEXPIRETIME).
- Hoist settings get/set into db::settings (&SqlitePool) and route both the
  Tauri settings commands and the MCP control layer through it, removing the
  third copy of the same SQL.
