+++ src-tauri/src/commands/settings.rs
@@ -4,30 +4,26 @@ use std::collections::HashMap;
 use tauri::State;
 
 #[tauri::command]
-pub async fn get_setting(pool: State<'_, SqlitePool>, key: String) -> Result<Option<String>, String> {
-    let result: Option<Setting> = sqlx::query_as("SELECT key, value FROM settings WHERE key = ?")
-        .bind(&key)
-        .fetch_optional(pool.inner())
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(result.map(|s| s.value))
+pub async fn get_setting(
+    pool: State<'_, SqlitePool>,
+    key: String,
+) -> Result<Option<String>, String> {
+    crate::db::settings::get(pool.inner(), &key).await
 }
 
 #[tauri::command]
-pub async fn set_setting(pool: State<'_, SqlitePool>, key: String, value: String) -> Result<(), String> {
-    sqlx::query("INSERT OR REPLACE INTO settings (key, value) VALUES (?, ?)")
-        .bind(&key)
-        .bind(&value)
-        .execute(pool.inner())
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(())
+pub async fn set_setting(
+    pool: State<'_, SqlitePool>,
+    key: String,
+    value: String,
+) -> Result<(), String> {
+    crate::db::settings::set(pool.inner(), &key, &value).await
 }
 
 #[tauri::command]
-pub async fn get_all_settings(pool: State<'_, SqlitePool>) -> Result<HashMap<String, String>, String> {
+pub async fn get_all_settings(
+    pool: State<'_, SqlitePool>,
+) -> Result<HashMap<String, String>, String> {
     let settings: Vec<Setting> = sqlx::query_as("SELECT key, value FROM settings")
         .fetch_all(pool.inner())
         .await
